requirements.txt
flask==2.3.3
flask-session==0.5.0
flask-login==0.6.3
scikit-learn==1.5.0
pandas==2.1.3
numpy>=1.26.0
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0
joblib==1.3.2
python-dotenv==1.0.0
mysql-connector-python==8.1.0
bcrypt==4.0.1
pyttsx3==2.90
fuzzywuzzy==0.18.0
requests==2.31.0
beautifulsoup4==4.12.2
